ai_chatbot/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
numpy